refactor: split orchestration store modules [OPE-486]
